+++ go/cmd/gvel/entity/account.go
@@ -17,31 +17,3 @@ type CreateAccountOutput struct {
 	PrivateKey    string
 	IsDefault     bool
 }
-
-type SetDefaultAccountInput struct {
-	Account string
-}
-
-type SetDefaultAccountOutput struct {
-	Account string
-}
-
-type ImportAccountInput struct {
-	PrivateKey          string
-	Passphrase          string
-	SetAsDefaultAccount bool
-}
-
-type ImportAccountOutput struct {
-	ImportedAccountAddress string
-}
-
-type ExportAccountInput struct {
-	PublicAddress string
-	Passphrase    string
-}
-
-type ExportAccountOutput struct {
-	PublicAddress string
-	PrivateKey    string
-}
