+++ go/cmd/gvel/entity/import_account.go
@@ -1,0 +1,11 @@
+package entity
+
+type ImportAccountInput struct {
+	PrivateKey          string
+	Passphrase          string
+	SetAsDefaultAccount bool
+}
+
+type ImportAccountOutput struct {
+	ImportedAccountAddress string
+}
