+++ go/cmd/gvel/entity/set_default_account.go
@@ -1,0 +1,9 @@
+package entity
+
+type SetDefaultAccountInput struct {
+	Account string
+}
+
+type SetDefaultAccountOutput struct {
+	Account string
+}
