+++ go/cmd/gvel/layers/commands/account/testhelper_test.go
@@ -26,7 +26,7 @@ type helper struct {
 	listCmd    *cobra.Command
 	defaultCmd *cobra.Command
 	importCmd  *cobra.Command
-	exportCmd  *cobra.Command
+	//exportCmd  *cobra.Command
 }
 
 func initTest(t *testing.T) *helper {
@@ -66,8 +66,7 @@ func initTest(t *testing.T) *helper {
 		// must be order by alphabet
 		createCmd:  cmd.Commands()[0],
 		defaultCmd: cmd.Commands()[1],
-		exportCmd:  cmd.Commands()[2],
-		importCmd:  cmd.Commands()[3],
-		listCmd:    cmd.Commands()[4],
+		importCmd:  cmd.Commands()[2],
+		listCmd:    cmd.Commands()[3],
 	}
 }
