--- go/cmd/gvel/layers/logic/export_account.go
@@ -1,32 +1,0 @@
-package logic
-
-import (
-	"encoding/json"
-	"github.com/pkg/errors"
-	"github.com/velo-protocol/DRSv2/go/cmd/gvel/entity"
-	"github.com/velo-protocol/DRSv2/go/cmd/gvel/utils/crypto"
-	"strings"
-)
-
-func (lo *logic) ExportAccount(input *entity.ExportAccountInput) (*entity.ExportAccountOutput, error) {
-	accountBytes, err := lo.DB.Get([]byte(input.PublicAddress))
-	if err != nil && !strings.Contains(err.Error(), "not found") {
-		return nil, errors.Wrapf(err, "account not found in config file, please run gvel account create or gvel account import")
-	}
-
-	var account entity.Account
-	err = json.Unmarshal(accountBytes, &account)
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to unmarshal account")
-	}
-
-	privKeyBytes, err := crypto.Decrypt(account.EncryptedPrivateKey, input.Passphrase)
-	if err != nil {
-		return nil, errors.Errorf("failed to decrypt the seed of %s with given passphrase: failed to decipher and authenticate: cipher: message authentication failed", input.PublicAddress)
-	}
-
-	return &entity.ExportAccountOutput{
-		PublicAddress: account.PublicAddress,
-		PrivateKey:    string(privKeyBytes),
-	}, nil
-}
