+++ go/cmd/gvel/layers/logic/interface.go
@@ -9,7 +9,6 @@ type Logic interface {
 	// Account module
 	CreateAccount(input *entity.CreateAccountInput) (*entity.CreateAccountOutput, error)
 	ImportAccount(input *entity.ImportAccountInput) (*entity.ImportAccountOutput, error)
-	ExportAccount(input *entity.ExportAccountInput) (*entity.ExportAccountOutput, error)
 	ListAccount() ([]*entity.Account, error)
 	SetDefaultAccount(input *entity.SetDefaultAccountInput) (*entity.SetDefaultAccountOutput, error)
